The primary key type has been changed in entities and repositories. Add `@Setter` annotations to auditing fields in `Author`, `Book`, and `BookGenre` models.

+++ building-api-training/library-api/src/main/java/com/gustavosantos/library_api/repository/AuthorRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.UUID;
 
 @Repository
-public interface AuthorRepository extends JpaRepository<Author, Long> {
+public interface AuthorRepository extends JpaRepository<Author, Integer> {
 
     Author findByPublicId(UUID publicId);
