+++ building-api-training/library-api/src/main/java/com/gustavosantos/library_api/repository/BookRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BookRepository extends JpaRepository<Book, Long> {
+public interface BookRepository extends JpaRepository<Book, Integer> {
 
     Book findByIsbn(String isbn);
